refactor: DeviceSection 로직 useDeviceSection으로 분리 (#75)
- 조회·매핑·연결/해제/이름변경·모달 로직을 useDeviceSection 훅으로 추출
- snake→camel 매핑을 훅에 가둬 컴포넌트는 UI 값만 사용
- DeviceSection은 로딩/에러 분기 + JSX만 남겨 순수 뷰로 정리
